refactor(server): hoist all state into AppContext; lifespan becomes per-session reference handout
