+++ iterm_mcpy/app_context.py
@@ -10,12 +10,135 @@ AppContext implements the read-only mapping protocol so existing tool code
 """
 
 import asyncio
+import dataclasses
 import logging
+import os
 from dataclasses import dataclass
-from typing import Any, Optional
+from datetime import datetime
+from typing import Any, Dict, List, Optional
+
+import iterm2
+
+from core.layouts import LayoutManager
+from core.terminal import ItermTerminal
+from core.agents import AgentRegistry
+from core.feedback import (
+    FeedbackHookManager,
+    FeedbackRegistry,
+    FeedbackForker,
+    GitHubIntegration,
+)
+from core.flows import get_event_bus, get_flow_manager
+from core.manager import ManagerRegistry
+from core.memory import SQLiteMemoryStore
+from core.models import AgentNotification
+from core.profiles import get_profile_manager
+from core.roles import RoleManager
+from core.service_hooks import get_service_hook_manager
+from core.services import get_service_manager
+from core.tags import SessionTagLockManager, FocusCooldownManager
+from utils.otel import init_tracing, shutdown_tracing
+from utils.telemetry import TelemetryEmitter
 
 logger = logging.getLogger("iterm-mcp-server")
 
+
+# ============================================================================
+# NOTIFICATION MANAGER
+# ============================================================================
+
+class NotificationManager:
+    """Manages agent notifications with ring buffer storage."""
+
+    # Status icons for compact display
+    STATUS_ICONS = {
+        "info": "ℹ",
+        "warning": "⚠",
+        "error": "✗",
+        "success": "✓",
+        "blocked": "⏸",
+    }
+
+    def __init__(self, max_per_agent: int = 50, max_total: int = 200):
+        self._notifications: List[AgentNotification] = []
+        self._max_per_agent = max_per_agent
+        self._max_total = max_total
+        self._lock = asyncio.Lock()
+
+    async def add(self, notification: AgentNotification) -> None:
+        """Add a notification, maintaining ring buffer limits."""
+        async with self._lock:
+            self._notifications.append(notification)
+            # Trim to max total
+            if len(self._notifications) > self._max_total:
+                self._notifications = self._notifications[-self._max_total:]
+
+    async def add_simple(
+        self,
+        agent: str,
+        level: str,
+        summary: str,
+        context: Optional[str] = None,
+        action_hint: Optional[str] = None,
+    ) -> None:
+        """Convenience method to add a notification."""
+        notification = AgentNotification(
+            agent=agent,
+            timestamp=datetime.now(),
+            level=level,  # type: ignore
+            summary=summary[:100],
+            context=context,
+            action_hint=action_hint,
+        )
+        await self.add(notification)
+
+    async def get(
+        self,
+        limit: int = 10,
+        level: Optional[str] = None,
+        agent: Optional[str] = None,
+        since: Optional[datetime] = None,
+    ) -> List[AgentNotification]:
+        """Get notifications with optional filters."""
+        async with self._lock:
+            result = self._notifications.copy()
+
+        # Apply filters
+        if level:
+            result = [n for n in result if n.level == level]
+        if agent:
+            result = [n for n in result if n.agent == agent]
+        if since:
+            result = [n for n in result if n.timestamp >= since]
+
+        # Return most recent first, limited
+        return sorted(result, key=lambda n: n.timestamp, reverse=True)[:limit]
+
+    async def get_latest_per_agent(self) -> Dict[str, AgentNotification]:
+        """Get the most recent notification for each agent."""
+        async with self._lock:
+            latest: Dict[str, AgentNotification] = {}
+            for n in reversed(self._notifications):
+                if n.agent not in latest:
+                    latest[n.agent] = n
+            return latest
+
+    def format_compact(self, notifications: List[AgentNotification]) -> str:
+        """Format notifications for compact TUI display."""
+        if not notifications:
+            return "━━━ No notifications ━━━"
+
+        lines = ["━━━ Agent Status ━━━"]
+        for n in notifications:
+            icon = self.STATUS_ICONS.get(n.level, "?")
+            lines.append(f"{n.agent:<12} {icon} {n.summary}")
+        lines.append("━━━━━━━━━━━━━━━━━━━━")
+        return "\n".join(lines)
+
+
+# ============================================================================
+# APP CONTEXT
+# ============================================================================
 
 @dataclass
 class AppContext:
@@ -53,16 +176,155 @@ class AppContext:
         return getattr(self, key, default)
 
     def __contains__(self, key: str) -> bool:
-        return hasattr(self, key)
+        return key in _FIELD_NAMES
 
+
+# Field-name set for __contains__ — avoids dunder false-positives from hasattr
+_FIELD_NAMES = {f.name for f in dataclasses.fields(AppContext)}
+
+
+# ============================================================================
+# SINGLETON
+# ============================================================================
 
 _app_context: Optional[AppContext] = None
 _init_lock = asyncio.Lock()
 
 
 async def _build_app_context() -> AppContext:
-    """Build the real context. Body moves here from iterm_lifespan in Task 2."""
-    raise NotImplementedError("populated in Task 2")
+    """Build the process-wide AppContext.
+
+    This runs exactly once per process lifetime (subsequent calls return the
+    cached singleton via get_app_context). Exceptions propagate — callers
+    handle failure.
+    """
+    # Configure logging
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+        handlers=[
+            logging.FileHandler(os.path.expanduser("~/.iterm-mcp.log")),
+            logging.StreamHandler()
+        ]
+    )
+    build_logger = logging.getLogger("iterm-mcp-server")
+    build_logger.info("Initializing iTerm2 connection...")
+
+    # Initialize OpenTelemetry tracing
+    tracing_enabled = init_tracing()
+    if tracing_enabled:
+        build_logger.info("OpenTelemetry tracing initialized successfully")
+    else:
+        build_logger.info(
+            "OpenTelemetry tracing not available "
+            "(install with: pip install iterm-mcp[otel])"
+        )
+
+    # Initialize iTerm2 connection
+    try:
+        connection = await iterm2.Connection.async_create()
+        build_logger.info("iTerm2 connection established successfully")
+    except Exception as conn_error:
+        build_logger.error(f"Failed to establish iTerm2 connection: {str(conn_error)}")
+        raise
+
+    # Initialize terminal controller
+    log_dir = os.path.expanduser("~/.iterm_mcp_logs")
+    terminal = ItermTerminal(
+        connection=connection,
+        log_dir=log_dir,
+        enable_logging=True,
+        default_max_lines=100,
+        max_snapshot_lines=1000
+    )
+    try:
+        await terminal.initialize()
+        build_logger.info("iTerm terminal controller initialized successfully")
+    except Exception as term_error:
+        build_logger.error(f"Failed to initialize iTerm terminal controller: {str(term_error)}")
+        raise
+
+    # Initialize layout manager
+    layout_manager = LayoutManager(terminal)
+
+    # Initialize agent registry
+    lock_manager = SessionTagLockManager()
+    agent_registry = AgentRegistry(lock_manager=lock_manager)
+
+    # Initialize telemetry emitter
+    telemetry = TelemetryEmitter(
+        log_manager=getattr(terminal, "log_manager", None),
+        agent_registry=agent_registry,
+    )
+
+    # Initialize notification manager
+    notification_manager = NotificationManager()
+
+    # Initialize focus cooldown manager
+    focus_cooldown = FocusCooldownManager()
+
+    # Initialize feedback system
+    feedback_registry = FeedbackRegistry()
+    feedback_hook_manager = FeedbackHookManager()
+    feedback_forker = FeedbackForker()
+    github_integration = GitHubIntegration()
+
+    # Initialize profile manager
+    profile_manager = get_profile_manager(build_logger)
+    build_logger.info(
+        f"Profile manager initialized with "
+        f"{len(profile_manager.list_team_profiles())} team profiles"
+    )
+
+    # Initialize service manager and hooks
+    service_manager = get_service_manager(logger=build_logger)
+    service_manager.set_terminal(terminal)
+    service_manager.load_global_config()
+    service_hook_manager = get_service_hook_manager(service_manager, build_logger)
+
+    # Initialize manager registry for hierarchical task delegation
+    manager_registry = ManagerRegistry()
+
+    # Initialize event bus and flow manager
+    event_bus = get_event_bus()
+    flow_manager = get_flow_manager()
+    await event_bus.start()
+
+    # Initialize role manager
+    role_manager = RoleManager(agent_registry=agent_registry)
+    build_logger.info(
+        f"Role manager initialized with "
+        f"{len(role_manager.list_roles())} role assignments"
+    )
+
+    # Initialize memory store
+    memory_store = SQLiteMemoryStore()
+    build_logger.info("Memory store initialized (SQLite with FTS5)")
+
+    return AppContext(
+        connection=connection,
+        terminal=terminal,
+        layout_manager=layout_manager,
+        agent_registry=agent_registry,
+        telemetry=telemetry,
+        notification_manager=notification_manager,
+        tag_lock_manager=lock_manager,
+        focus_cooldown=focus_cooldown,
+        feedback_registry=feedback_registry,
+        feedback_hook_manager=feedback_hook_manager,
+        feedback_forker=feedback_forker,
+        github_integration=github_integration,
+        profile_manager=profile_manager,
+        service_manager=service_manager,
+        service_hook_manager=service_hook_manager,
+        manager_registry=manager_registry,
+        event_bus=event_bus,
+        flow_manager=flow_manager,
+        role_manager=role_manager,
+        memory_store=memory_store,
+        logger=build_logger,
+        log_dir=log_dir,
+    )
 
 
 async def get_app_context() -> AppContext:
@@ -92,3 +354,4 @@ async def shutdown_app_context() -> None:
             await ctx.event_bus.stop()
         except Exception:
             logger.exception("Error stopping event bus during shutdown")
+    shutdown_tracing()
